requirements.txt
Flask>=2.0.2
Werkzeug>=2.0.0
Flask-JWT-Extended>=4.3.1
PyJWT>=2.1.0
Flask-SQLAlchemy>=2.5.1
python-dotenv==0.19.0
Flask-Migrate
